Rewrite of DungeonDataMgr
Implemented dungeonDataMgr functions
- handling in-memory dungeon data
- Updating dungeons
- Removed old structures

DungeonDataMgr update
- DungeonDataMgr now compiles
- Dungeons can be loaded, updated
