+++ src/eve-common/EVE_Dungeon.h
@@ -145,19 +145,6 @@ namespace Dungeon {
             FacWarLoyaltyPointsGranted = 48
         };
     }
-
-    /* POD structure for active dungeon */
-    struct ActiveData {
-        uint8 state;
-        uint32 systemID;
-        uint32 dunItemID;
-        uint32 dunTemplateID;
-        int64 dunExpiryTime;
-        double x;
-        double y;
-        double z;
-    };
-    
     /* POD structure entries for dungeon data */
     struct Template {
         uint8 dunTypeID;
@@ -169,13 +156,9 @@ namespace Dungeon {
         uint32 dunFactionID;
     };
 
-    struct RoomInfo {
-        uint8 dunRoomType;
-        uint8 dunRoomCategory;
-        uint8 dunRoomSpawnID;
-        uint8 dunRoomSpawnType;
-        uint16 dunRoomID;
-    };
+    /* Tags for dungeon multi-index container */
+    struct DungeonsByID {};
+    struct DungeonsByArchetype {};
 
     struct RoomObject {
         uint32 objectID;
@@ -191,39 +174,23 @@ namespace Dungeon {
         double radius;
     };
 
-    struct RoomData {
+    struct Room {
+        uint16 roomID;
+        std::string roomName;
+        std::vector<RoomObject> objects;
         int16 x;
         int16 y;
         int16 z;
-        int32 dunGroupID;
     };
 
-    struct GroupData {
-        uint8 typeCatID;    // this is categoryID of the itemType, and needed to simplify create/spawn code
-        int16 typeID;
-        int16 typeGrpID;   // this is groupID of the itemType, and needed to simplify create/spawn code
-        int16 x;
-        int16 y;
-        int16 z;
-        uint16 radius;
-        std::string typeName;
+    struct Dungeon {
+        uint32 dungeonID;
+        std::string name;
+        uint8 status;
+        uint32 factionID;
+        uint8 archetypeID;
+        std::map<uint16, Room> rooms;
     };
-
-    struct RoomSpawnInfo {
-        int16 x;
-        int16 y;
-        int16 z;
-        uint16 dunRoomSpawnID;
-        uint16 dunRoomSpawnType;
-    };
-
-    struct EntryData {
-        int16 x;
-        int16 y;
-        int16 z;
-        uint16 dunEntryID;
-    };
-
 }
 
 /*
